test(multiorch): move ShardAnalysis helper tests out of generator_test
The generator no longer computes leader verdicts — it only assembles a
ShardAnalysis (leader identity, grouping, Now/Policy). Tests that exercise the
call-time predicate helpers (leaderServing, leaderPostgresReady,
replicasStreamingFromLeader, IsInStandbyList) were still living in
generator_test.go under stale TestPopulatePrimaryInfo_* names referencing a
function that no longer exists.

Move those into shard_analysis_helpers_test.go, colocated with the helpers,
and rename them for what they test. Drop the redundant
TestShardAnalysis_IsInStandbyList (the table-driven TestIsInStandbyList already
subsumes it) and trim two incidental leaderServing assertions from
generator_test.go's identity tests. No production changes.
