+++ go/services/multiorch/recovery/analysis/generator_test.go
@@ -23,8 +23,6 @@ import (
 	"google.golang.org/protobuf/types/known/durationpb"
 	"google.golang.org/protobuf/types/known/timestamppb"
 
-	"github.com/multigres/multigres/go/common/topoclient"
-
 	commonconsensus "github.com/multigres/multigres/go/common/consensus"
 	clustermetadatapb "github.com/multigres/multigres/go/pb/clustermetadata"
 	multiorchdatapb "github.com/multigres/multigres/go/pb/multiorchdata"
@@ -286,7 +284,6 @@ func TestAnalysisGenerator_GenerateShardAnalyses_Replica(t *testing.T) {
 
 	// Primary health is now a shard-level field
 	assert.NotNil(t, sa.HighestShardRule.GetLeaderId(), "should have topology primary ID populated")
-	assert.True(t, leaderServing(sa))
 }
 
 func TestAnalysisGenerator_GenerateShardAnalyses_MultipleTableGroups(t *testing.T) {
@@ -417,9 +414,8 @@ func TestPopulatePrimaryInfo_NoPrimaryInShard(t *testing.T) {
 	sa, err := gen.GenerateShardAnalysis(&clustermetadatapb.ShardKey{Database: "db1", TableGroup: "tg1", Shard: "shard1"})
 	require.NoError(t, err)
 
-	// When no primary exists in the shard, topology primary fields should be nil/false
+	// When no primary exists in the shard, topology primary fields should be nil.
 	assert.Nil(t, sa.HighestShardRule.GetLeaderId())
-	assert.False(t, leaderServing(sa))
 }
 
 // Task 7: Test for primary with postgres down
@@ -705,1084 +701,6 @@ func TestGenerateShardAnalysis_StaleLeaderSupersededViaFollowerRule(t *testing.T
 		"the leader named by the follower's higher-term rule must win over the stale self-claiming leader")
 	assert.Equal(t, int64(6), sa.HighestShardRule.GetRuleNumber().GetCoordinatorTerm())
 	assert.Nil(t, sa.Leader, "the new leader has no health state yet, so Leader is nil")
-}
-
-func TestIsInStandbyList(t *testing.T) {
-	ps := store.NewTestCache(t)
-
-	primaryID := &clustermetadatapb.ID{
-		Component: clustermetadatapb.ID_MULTIPOOLER,
-		Cell:      "cell1",
-		Name:      "primary-1",
-	}
-
-	replica1ID := &clustermetadatapb.ID{
-		Component: clustermetadatapb.ID_MULTIPOOLER,
-		Cell:      "cell1",
-		Name:      "replica-1",
-	}
-
-	replica2ID := &clustermetadatapb.ID{
-		Component: clustermetadatapb.ID_MULTIPOOLER,
-		Cell:      "cell2",
-		Name:      "replica-2",
-	}
-
-	tests := []struct {
-		name          string
-		replicaID     *clustermetadatapb.ID
-		primaryStatus *multipoolermanagerdatapb.PrimaryStatus
-		expected      bool
-	}{
-		{
-			name:      "replica in standby list",
-			replicaID: replica1ID,
-			primaryStatus: &multipoolermanagerdatapb.PrimaryStatus{
-				Lsn:   "0/1234567",
-				Ready: true,
-				SyncReplicationConfig: &multipoolermanagerdatapb.SynchronousReplicationConfiguration{
-					StandbyIds: []*clustermetadatapb.ID{replica1ID},
-				},
-			},
-			expected: true,
-		},
-		{
-			name:      "replica not in standby list",
-			replicaID: replica2ID,
-			primaryStatus: &multipoolermanagerdatapb.PrimaryStatus{
-				Lsn:   "0/1234567",
-				Ready: true,
-				SyncReplicationConfig: &multipoolermanagerdatapb.SynchronousReplicationConfiguration{
-					StandbyIds: []*clustermetadatapb.ID{replica1ID},
-				},
-			},
-			expected: false,
-		},
-		{
-			name:      "empty standby list",
-			replicaID: replica1ID,
-			primaryStatus: &multipoolermanagerdatapb.PrimaryStatus{
-				Lsn:   "0/1234567",
-				Ready: true,
-				SyncReplicationConfig: &multipoolermanagerdatapb.SynchronousReplicationConfiguration{
-					StandbyIds: []*clustermetadatapb.ID{},
-				},
-			},
-			expected: false,
-		},
-		{
-			name:      "nil sync replication config",
-			replicaID: replica1ID,
-			primaryStatus: &multipoolermanagerdatapb.PrimaryStatus{
-				Lsn:                   "0/1234567",
-				Ready:                 true,
-				SyncReplicationConfig: nil,
-			},
-			expected: false,
-		},
-		{
-			name:          "nil primary status",
-			replicaID:     replica1ID,
-			primaryStatus: nil,
-			expected:      false,
-		},
-		{
-			name:      "multiple standbys - replica present",
-			replicaID: replica2ID,
-			primaryStatus: &multipoolermanagerdatapb.PrimaryStatus{
-				Lsn:   "0/1234567",
-				Ready: true,
-				SyncReplicationConfig: &multipoolermanagerdatapb.SynchronousReplicationConfiguration{
-					StandbyIds: []*clustermetadatapb.ID{replica1ID, replica2ID},
-				},
-			},
-			expected: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Set up pooler store with primary
-			store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-				MultiPooler: &clustermetadatapb.MultiPooler{
-					Id: primaryID,
-					ShardKey: &clustermetadatapb.ShardKey{
-						Database:   "testdb",
-						TableGroup: "testtg",
-						Shard:      "0",
-					},
-					Type: clustermetadatapb.PoolerType_PRIMARY,
-				},
-				IsLastCheckValid: true,
-				LastSeen:         timestamppb.Now(),
-				ConsensusStatus: &clustermetadatapb.ConsensusStatus{
-					Id: primaryID,
-					CurrentPosition: &clustermetadatapb.PoolerPosition{
-						Rule: &clustermetadatapb.ShardRule{
-							RuleNumber:    &clustermetadatapb.RuleNumber{CoordinatorTerm: 1},
-							LeaderId:      primaryID,
-							CohortMembers: tt.primaryStatus.GetSyncReplicationConfig().GetStandbyIds(),
-						},
-					},
-				},
-				Status: &multipoolermanagerdatapb.Status{
-					PoolerType:    clustermetadatapb.PoolerType_PRIMARY,
-					PostgresReady: true,
-					PrimaryStatus: tt.primaryStatus,
-				},
-			}, nil))
-
-			generator := NewAnalysisGenerator(ps, nil)
-			sa, err := generator.GenerateShardAnalysis(&clustermetadatapb.ShardKey{Database: "testdb", TableGroup: "testtg", Shard: "0"})
-			require.NoError(t, err)
-
-			result := sa.IsInStandbyList(tt.replicaID)
-
-			assert.Equal(t, tt.expected, result)
-		})
-	}
-}
-
-func TestPopulatePrimaryInfo_PrimaryHealthFields(t *testing.T) {
-	t.Run("sets PrimaryPoolerReachable and PrimaryPostgresReady correctly", func(t *testing.T) {
-		ps := store.NewTestCache(t)
-
-		// Primary with pooler reachable and postgres running
-		respondedAt := time.Now().Add(-3 * time.Second)
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{
-					Component: clustermetadatapb.ID_MULTIPOOLER,
-					Cell:      "cell1",
-					Name:      "primary",
-				},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-				Hostname: "primary-host",
-				PortMap:  map[string]int32{"postgres": 5432},
-			},
-			ConsensusStatus:       primaryConsensusStatus(&clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "primary"}, 1),
-			IsLastCheckValid:      true,
-			LastSeen:              timestamppb.Now(),
-			LastPostgresReadyTime: timestamppb.New(respondedAt),
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType:    clustermetadatapb.PoolerType_PRIMARY,
-				PostgresReady: true,
-			},
-		}, nil))
-
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{
-					Component: clustermetadatapb.ID_MULTIPOOLER,
-					Cell:      "cell1",
-					Name:      "replica",
-				},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-			},
-			IsLastCheckValid: true,
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType: clustermetadatapb.PoolerType_REPLICA,
-			},
-		}, nil))
-
-		gen := NewAnalysisGenerator(ps, nil)
-		sa, err := gen.GenerateShardAnalysis(&clustermetadatapb.ShardKey{Database: "db1", TableGroup: "tg1", Shard: "shard1"})
-		require.NoError(t, err)
-
-		assert.True(t, leaderPostgresReady(sa))
-		assert.True(t, leaderServing(sa))
-	})
-
-	t.Run("sets PrimaryPoolerReachable false when pooler unreachable", func(t *testing.T) {
-		ps := store.NewTestCache(t)
-
-		// Primary with pooler unreachable
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{
-					Component: clustermetadatapb.ID_MULTIPOOLER,
-					Cell:      "cell1",
-					Name:      "primary",
-				},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-				Hostname: "primary-host",
-				PortMap:  map[string]int32{"postgres": 5432},
-			},
-			ConsensusStatus:  primaryConsensusStatus(&clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "primary"}, 1),
-			IsLastCheckValid: false, // Pooler unreachable
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType:    clustermetadatapb.PoolerType_PRIMARY,
-				PostgresReady: false,
-			},
-		}, nil))
-
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{
-					Component: clustermetadatapb.ID_MULTIPOOLER,
-					Cell:      "cell1",
-					Name:      "replica",
-				},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-			},
-			IsLastCheckValid: true,
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType: clustermetadatapb.PoolerType_REPLICA,
-			},
-		}, nil))
-
-		gen := NewAnalysisGenerator(ps, nil)
-		sa, err := gen.GenerateShardAnalysis(&clustermetadatapb.ShardKey{Database: "db1", TableGroup: "tg1", Shard: "shard1"})
-		require.NoError(t, err)
-
-		assert.False(t, leaderPostgresReady(sa))
-		assert.False(t, leaderServing(sa))
-	})
-}
-
-func TestAllReplicasConnectedToLeader(t *testing.T) {
-	t.Run("returns true when all replicas connected", func(t *testing.T) {
-		ps := store.NewTestCache(t)
-
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{
-					Component: clustermetadatapb.ID_MULTIPOOLER,
-					Cell:      "cell1",
-					Name:      "primary",
-				},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-				Hostname: "primary-host",
-				PortMap:  map[string]int32{"postgres": 5432},
-			},
-			ConsensusStatus:  primaryConsensusStatus(&clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "primary"}, 1),
-			IsLastCheckValid: false, // Primary pooler is down
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType:    clustermetadatapb.PoolerType_PRIMARY,
-				PostgresReady: false,
-			},
-		}, nil))
-
-		now := time.Now()
-
-		// Replica 1 - connected to primary
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{
-					Component: clustermetadatapb.ID_MULTIPOOLER,
-					Cell:      "cell1",
-					Name:      "replica1",
-				},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-			},
-			IsLastCheckValid:        true,
-			LastSeen:                timestamppb.New(time.Now()),
-			StreamSnapshotsReceived: 1,
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType: clustermetadatapb.PoolerType_REPLICA,
-				ReplicationStatus: &multipoolermanagerdatapb.StandbyReplicationStatus{
-					LastReceiveLsn:     "0/1234567",
-					WalReceiverStatus:  "streaming",
-					LastMsgReceiveTime: timestamppb.New(now.Add(-5 * time.Second)),
-					PrimaryConnInfo: &multipoolermanagerdatapb.PrimaryConnInfo{
-						Host: "primary-host",
-						Port: 5432,
-					},
-				},
-			},
-		}, nil))
-
-		// Replica 2 - also connected to primary
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{
-					Component: clustermetadatapb.ID_MULTIPOOLER,
-					Cell:      "cell1",
-					Name:      "replica2",
-				},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-			},
-			IsLastCheckValid:        true,
-			LastSeen:                timestamppb.New(time.Now()),
-			StreamSnapshotsReceived: 1,
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType: clustermetadatapb.PoolerType_REPLICA,
-				ReplicationStatus: &multipoolermanagerdatapb.StandbyReplicationStatus{
-					LastReceiveLsn:     "0/1234567",
-					WalReceiverStatus:  "streaming",
-					LastMsgReceiveTime: timestamppb.New(now.Add(-5 * time.Second)),
-					PrimaryConnInfo: &multipoolermanagerdatapb.PrimaryConnInfo{
-						Host: "primary-host",
-						Port: 5432,
-					},
-				},
-			},
-		}, nil))
-
-		gen := NewAnalysisGenerator(ps, nil)
-		sa, err := gen.GenerateShardAnalysis(&clustermetadatapb.ShardKey{Database: "db1", TableGroup: "tg1", Shard: "shard1"})
-		require.NoError(t, err)
-
-		assert.True(t, replicasStreamingFromLeader(sa), "should be true when all replicas are connected")
-	})
-
-	t.Run("returns false when one replica disconnected", func(t *testing.T) {
-		ps := store.NewTestCache(t)
-
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{
-					Component: clustermetadatapb.ID_MULTIPOOLER,
-					Cell:      "cell1",
-					Name:      "primary",
-				},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-				Hostname: "primary-host",
-				PortMap:  map[string]int32{"postgres": 5432},
-			},
-			ConsensusStatus:  primaryConsensusStatus(&clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "primary"}, 1),
-			IsLastCheckValid: false,
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType:    clustermetadatapb.PoolerType_PRIMARY,
-				PostgresReady: false,
-			},
-		}, nil))
-
-		// Replica 1 - connected
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{
-					Component: clustermetadatapb.ID_MULTIPOOLER,
-					Cell:      "cell1",
-					Name:      "replica1",
-				},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-			},
-			IsLastCheckValid:        true,
-			LastSeen:                timestamppb.New(time.Now()),
-			StreamSnapshotsReceived: 1,
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType: clustermetadatapb.PoolerType_REPLICA,
-				ReplicationStatus: &multipoolermanagerdatapb.StandbyReplicationStatus{
-					LastReceiveLsn:     "0/1234567",
-					WalReceiverStatus:  "streaming",
-					LastMsgReceiveTime: timestamppb.New(time.Now().Add(-5 * time.Second)),
-					PrimaryConnInfo: &multipoolermanagerdatapb.PrimaryConnInfo{
-						Host: "primary-host",
-						Port: 5432,
-					},
-				},
-			},
-		}, nil))
-
-		// Replica 2 - disconnected (no PrimaryConnInfo)
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{
-					Component: clustermetadatapb.ID_MULTIPOOLER,
-					Cell:      "cell1",
-					Name:      "replica2",
-				},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-			},
-			IsLastCheckValid:        true,
-			LastSeen:                timestamppb.New(time.Now()),
-			StreamSnapshotsReceived: 1,
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType:        clustermetadatapb.PoolerType_REPLICA,
-				ReplicationStatus: &multipoolermanagerdatapb.StandbyReplicationStatus{
-					// No PrimaryConnInfo - replica is disconnected
-				},
-			},
-		}, nil))
-
-		gen := NewAnalysisGenerator(ps, nil)
-		sa, err := gen.GenerateShardAnalysis(&clustermetadatapb.ShardKey{Database: "db1", TableGroup: "tg1", Shard: "shard1"})
-		require.NoError(t, err)
-
-		assert.False(t, replicasStreamingFromLeader(sa), "should be false when any replica is disconnected")
-	})
-
-	t.Run("returns false when replica unreachable", func(t *testing.T) {
-		ps := store.NewTestCache(t)
-
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{
-					Component: clustermetadatapb.ID_MULTIPOOLER,
-					Cell:      "cell1",
-					Name:      "primary",
-				},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-				Hostname: "primary-host",
-				PortMap:  map[string]int32{"postgres": 5432},
-			},
-			ConsensusStatus:  primaryConsensusStatus(&clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "primary"}, 1),
-			IsLastCheckValid: false,
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType:    clustermetadatapb.PoolerType_PRIMARY,
-				PostgresReady: false,
-			},
-		}, nil))
-
-		// Replica was previously healthy but is now unreachable (stream disconnected).
-		// StreamSnapshotsReceived > 0 distinguishes this from a brand-new pooler.
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{
-					Component: clustermetadatapb.ID_MULTIPOOLER,
-					Cell:      "cell1",
-					Name:      "replica1",
-				},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-			},
-			IsLastCheckValid:        false, // Replica unreachable
-			StreamSnapshotsReceived: 1,     // Was previously healthy
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType: clustermetadatapb.PoolerType_REPLICA,
-			},
-		}, nil))
-
-		gen := NewAnalysisGenerator(ps, nil)
-		sa, err := gen.GenerateShardAnalysis(&clustermetadatapb.ShardKey{Database: "db1", TableGroup: "tg1", Shard: "shard1"})
-		require.NoError(t, err)
-
-		assert.False(t, replicasStreamingFromLeader(sa), "should be false when replica is unreachable")
-	})
-
-	t.Run("returns false when no replicas exist", func(t *testing.T) {
-		ps := store.NewTestCache(t)
-
-		// Only primary, no replicas
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{
-					Component: clustermetadatapb.ID_MULTIPOOLER,
-					Cell:      "cell1",
-					Name:      "primary",
-				},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-				Hostname: "primary-host",
-				PortMap:  map[string]int32{"postgres": 5432},
-			},
-			ConsensusStatus:  primaryConsensusStatus(&clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "primary"}, 1),
-			IsLastCheckValid: true,
-			LastSeen:         timestamppb.New(time.Now()),
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType:    clustermetadatapb.PoolerType_PRIMARY,
-				PostgresReady: true,
-			},
-		}, nil))
-
-		gen := NewAnalysisGenerator(ps, nil)
-		sa, err := gen.GenerateShardAnalysis(&clustermetadatapb.ShardKey{Database: "db1", TableGroup: "tg1", Shard: "shard1"})
-		require.NoError(t, err)
-
-		// Primary-only shard: ReplicasConnectedToLeader should be false (no replicas)
-		assert.False(t, replicasStreamingFromLeader(sa))
-	})
-
-	t.Run("returns false when replica pointing to wrong primary", func(t *testing.T) {
-		ps := store.NewTestCache(t)
-
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{
-					Component: clustermetadatapb.ID_MULTIPOOLER,
-					Cell:      "cell1",
-					Name:      "primary",
-				},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-				Hostname: "primary-host",
-				PortMap:  map[string]int32{"postgres": 5432},
-			},
-			ConsensusStatus:  primaryConsensusStatus(&clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "primary"}, 1),
-			IsLastCheckValid: false,
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType:    clustermetadatapb.PoolerType_PRIMARY,
-				PostgresReady: false,
-			},
-		}, nil))
-
-		// Replica pointing to different host
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{
-					Component: clustermetadatapb.ID_MULTIPOOLER,
-					Cell:      "cell1",
-					Name:      "replica",
-				},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-			},
-			IsLastCheckValid:        true,
-			LastSeen:                timestamppb.New(time.Now()),
-			StreamSnapshotsReceived: 1,
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType: clustermetadatapb.PoolerType_REPLICA,
-				ReplicationStatus: &multipoolermanagerdatapb.StandbyReplicationStatus{
-					LastReceiveLsn:     "0/1234567",
-					WalReceiverStatus:  "streaming",
-					LastMsgReceiveTime: timestamppb.New(time.Now().Add(-5 * time.Second)),
-					PrimaryConnInfo: &multipoolermanagerdatapb.PrimaryConnInfo{
-						Host: "different-host", // Wrong host!
-						Port: 5432,
-					},
-				},
-			},
-		}, nil))
-
-		gen := NewAnalysisGenerator(ps, nil)
-		sa, err := gen.GenerateShardAnalysis(&clustermetadatapb.ShardKey{Database: "db1", TableGroup: "tg1", Shard: "shard1"})
-		require.NoError(t, err)
-
-		assert.False(t, replicasStreamingFromLeader(sa), "should be false when replica points to wrong primary")
-	})
-
-	t.Run("returns false when WAL receiver is not streaming", func(t *testing.T) {
-		ps := store.NewTestCache(t)
-
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "primary"},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-				Hostname: "primary-host",
-				PortMap:  map[string]int32{"postgres": 5432},
-			},
-			ConsensusStatus: primaryConsensusStatus(&clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "primary"}, 1),
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType: clustermetadatapb.PoolerType_PRIMARY,
-			},
-		}, nil))
-
-		for _, status := range []string{"", "starting", "waiting", "stopping"} {
-			store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-				MultiPooler: &clustermetadatapb.MultiPooler{
-					Id: &clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "replica"},
-					ShardKey: &clustermetadatapb.ShardKey{
-						Database:   "db1",
-						TableGroup: "tg1",
-						Shard:      "shard1",
-					},
-				},
-				IsLastCheckValid:        true,
-				LastSeen:                timestamppb.New(time.Now()),
-				StreamSnapshotsReceived: 1,
-				Status: &multipoolermanagerdatapb.Status{
-					PoolerType: clustermetadatapb.PoolerType_REPLICA,
-					ReplicationStatus: &multipoolermanagerdatapb.StandbyReplicationStatus{
-						LastReceiveLsn:     "0/1234567",
-						WalReceiverStatus:  status, // not "streaming"
-						LastMsgReceiveTime: timestamppb.New(time.Now().Add(-5 * time.Second)),
-						PrimaryConnInfo: &multipoolermanagerdatapb.PrimaryConnInfo{
-							Host: "primary-host",
-							Port: 5432,
-						},
-					},
-				},
-			}, nil))
-
-			gen := NewAnalysisGenerator(ps, nil)
-			analysis, err := gen.GenerateAnalysisForPooler(topoclient.ComponentID("multipooler-cell1-replica"))
-			require.NoError(t, err)
-			assert.False(t, replicasStreamingFromLeader(analysis), "should be false when wal_receiver_status=%q", status)
-		}
-	})
-
-	t.Run("returns false when last_msg_receive_time is stale (default threshold)", func(t *testing.T) {
-		// No WalReceiverStatusInterval supplied — falls back to defaultReplicationHeartbeatStalenessThreshold.
-		ps := store.NewTestCache(t)
-
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "primary"},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-				Hostname: "primary-host",
-				PortMap:  map[string]int32{"postgres": 5432},
-			},
-			ConsensusStatus: primaryConsensusStatus(&clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "primary"}, 1),
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType: clustermetadatapb.PoolerType_PRIMARY,
-			},
-		}, nil))
-
-		fixedNow := time.Date(2026, 1, 1, 12, 0, 0, 0, time.UTC)
-		staleTime := fixedNow.Add(-(defaultReplicationHeartbeatStalenessThreshold + time.Second))
-
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "replica"},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-			},
-			IsLastCheckValid:        true,
-			LastSeen:                timestamppb.New(time.Now()),
-			StreamSnapshotsReceived: 1,
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType: clustermetadatapb.PoolerType_REPLICA,
-				ReplicationStatus: &multipoolermanagerdatapb.StandbyReplicationStatus{
-					LastReceiveLsn:     "0/1234567",
-					WalReceiverStatus:  "streaming",
-					LastMsgReceiveTime: timestamppb.New(staleTime),
-					// WalReceiverStatusInterval intentionally nil — exercises fallback path
-					PrimaryConnInfo: &multipoolermanagerdatapb.PrimaryConnInfo{
-						Host: "primary-host",
-						Port: 5432,
-					},
-				},
-			},
-		}, nil))
-
-		gen := NewAnalysisGenerator(ps, nil)
-		gen.now = func() time.Time { return fixedNow }
-		analysis, err := gen.GenerateAnalysisForPooler(topoclient.ComponentID("multipooler-cell1-replica"))
-		require.NoError(t, err)
-
-		assert.False(t, replicasStreamingFromLeader(analysis), "should be false when last_msg_receive_time is stale")
-	})
-
-	t.Run("returns false when last_msg_receive_time is stale (dynamic threshold)", func(t *testing.T) {
-		// WalReceiverStatusInterval supplied — threshold is multiplier × interval.
-		ps := store.NewTestCache(t)
-
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "primary"},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-				Hostname: "primary-host",
-				PortMap:  map[string]int32{"postgres": 5432},
-			},
-			ConsensusStatus: primaryConsensusStatus(&clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "primary"}, 1),
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType: clustermetadatapb.PoolerType_PRIMARY,
-			},
-		}, nil))
-
-		fixedNow := time.Date(2026, 1, 1, 12, 0, 0, 0, time.UTC)
-		interval := 5 * time.Second
-		dynamicThreshold := replicationHeartbeatStalenessMultiplier * interval // 15s
-		staleTime := fixedNow.Add(-(dynamicThreshold + time.Second))
-
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "replica"},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-			},
-			IsLastCheckValid:        true,
-			LastSeen:                timestamppb.New(time.Now()),
-			StreamSnapshotsReceived: 1,
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType: clustermetadatapb.PoolerType_REPLICA,
-				ReplicationStatus: &multipoolermanagerdatapb.StandbyReplicationStatus{
-					LastReceiveLsn:            "0/1234567",
-					WalReceiverStatus:         "streaming",
-					LastMsgReceiveTime:        timestamppb.New(staleTime),
-					WalReceiverStatusInterval: durationpb.New(interval),
-					PrimaryConnInfo: &multipoolermanagerdatapb.PrimaryConnInfo{
-						Host: "primary-host",
-						Port: 5432,
-					},
-				},
-			},
-		}, nil))
-
-		gen := NewAnalysisGenerator(ps, nil)
-		gen.now = func() time.Time { return fixedNow }
-		analysis, err := gen.GenerateAnalysisForPooler(topoclient.ComponentID("multipooler-cell1-replica"))
-		require.NoError(t, err)
-
-		assert.False(t, replicasStreamingFromLeader(analysis), "should be false when last_msg_receive_time exceeds dynamic threshold")
-	})
-
-	t.Run("returns false when last_msg_receive_time exceeds wal_receiver_timeout", func(t *testing.T) {
-		// Even if the delay is below the staleness threshold, if it exceeds the
-		// WAL receiver timeout the connection is effectively dead.
-		ps := store.NewTestCache(t)
-
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "primary"},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-				Hostname: "primary-host",
-				PortMap:  map[string]int32{"postgres": 5432},
-			},
-			ConsensusStatus: primaryConsensusStatus(&clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "primary"}, 1),
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType: clustermetadatapb.PoolerType_PRIMARY,
-			},
-		}, nil))
-
-		fixedNow := time.Date(2026, 1, 1, 12, 0, 0, 0, time.UTC)
-		walReceiverTimeout := 60 * time.Second
-		// last_msg_receive_time is 61s ago — exceeds wal_receiver_timeout (60s) but
-		// is still within the staleness threshold (3×10s = 30s would be fine, but
-		// the hard deadline fires first).
-		lastMsgReceiveTime := fixedNow.Add(-(walReceiverTimeout + time.Second))
-
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "replica"},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-			},
-			IsLastCheckValid:        true,
-			LastSeen:                timestamppb.New(time.Now()),
-			StreamSnapshotsReceived: 1,
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType: clustermetadatapb.PoolerType_REPLICA,
-				ReplicationStatus: &multipoolermanagerdatapb.StandbyReplicationStatus{
-					LastReceiveLsn:            "0/1234567",
-					WalReceiverStatus:         "streaming",
-					LastMsgReceiveTime:        timestamppb.New(lastMsgReceiveTime),
-					WalReceiverStatusInterval: durationpb.New(10 * time.Second),
-					WalReceiverTimeout:        durationpb.New(walReceiverTimeout),
-					PrimaryConnInfo: &multipoolermanagerdatapb.PrimaryConnInfo{
-						Host: "primary-host",
-						Port: 5432,
-					},
-				},
-			},
-		}, nil))
-
-		gen := NewAnalysisGenerator(ps, nil)
-		gen.now = func() time.Time { return fixedNow }
-		analysis, err := gen.GenerateAnalysisForPooler(topoclient.ComponentID("multipooler-cell1-replica"))
-		require.NoError(t, err)
-
-		assert.False(t, replicasStreamingFromLeader(analysis), "should be false when delay exceeds wal_receiver_timeout")
-	})
-
-	t.Run("returns true when last_msg_receive_time is nil", func(t *testing.T) {
-		// Backward compatibility: replicas that don't report last_msg_receive_time
-		// (e.g. running an older version) should still be considered connected if
-		// the WAL receiver is streaming.
-		ps := store.NewTestCache(t)
-
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "primary"},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-				Hostname: "primary-host",
-				PortMap:  map[string]int32{"postgres": 5432},
-			},
-			ConsensusStatus: primaryConsensusStatus(&clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "primary"}, 1),
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType: clustermetadatapb.PoolerType_PRIMARY,
-			},
-		}, nil))
-
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id: &clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "replica"},
-				ShardKey: &clustermetadatapb.ShardKey{
-					Database:   "db1",
-					TableGroup: "tg1",
-					Shard:      "shard1",
-				},
-			},
-			IsLastCheckValid:        true,
-			LastSeen:                timestamppb.New(time.Now()),
-			StreamSnapshotsReceived: 1,
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType: clustermetadatapb.PoolerType_REPLICA,
-				ReplicationStatus: &multipoolermanagerdatapb.StandbyReplicationStatus{
-					LastReceiveLsn:    "0/1234567",
-					WalReceiverStatus: "streaming",
-					// LastMsgReceiveTime intentionally nil
-					PrimaryConnInfo: &multipoolermanagerdatapb.PrimaryConnInfo{
-						Host: "primary-host",
-						Port: 5432,
-					},
-				},
-			},
-		}, nil))
-
-		gen := NewAnalysisGenerator(ps, nil)
-		analysis, err := gen.GenerateAnalysisForPooler(topoclient.ComponentID("multipooler-cell1-replica"))
-		require.NoError(t, err)
-
-		assert.True(t, replicasStreamingFromLeader(analysis), "should be true when last_msg_receive_time is nil")
-	})
-
-	t.Run("new pooler with no health data does not count against connected replicas", func(t *testing.T) {
-		// Regression test for the startup race: PoolerWatcher adds a pooler to the
-		// store before its health stream delivers the first snapshot
-		// (StreamSnapshotsReceived==0). If the recovery tick fires in that window,
-		// the new pooler must not be counted as a "disconnected replica" —
-		// otherwise it would make allReplicasConnectedToLeader return false and
-		// disable the LeaderIsDeadAnalyzer suppression window, risking a
-		// premature failover.
-		ps := store.NewTestCache(t)
-
-		now := time.Now()
-
-		// Leader is transiently unreachable (pooler down, postgres still running).
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id:       &clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "primary"},
-				ShardKey: &clustermetadatapb.ShardKey{Database: "db1", TableGroup: "tg1", Shard: "shard1"},
-				Hostname: "primary-host",
-				PortMap:  map[string]int32{"postgres": 5432},
-			},
-			ConsensusStatus:         primaryConsensusStatus(&clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "primary"}, 1),
-			IsLastCheckValid:        false,
-			StreamSnapshotsReceived: 5,
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType:    clustermetadatapb.PoolerType_PRIMARY,
-				PostgresReady: false,
-			},
-		}, nil))
-
-		// Existing replica: healthy and actively streaming WAL from the leader.
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id:       &clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell1", Name: "replica1"},
-				ShardKey: &clustermetadatapb.ShardKey{Database: "db1", TableGroup: "tg1", Shard: "shard1"},
-			},
-			IsLastCheckValid:        true,
-			LastSeen:                timestamppb.New(time.Now()),
-			StreamSnapshotsReceived: 10,
-			Status: &multipoolermanagerdatapb.Status{
-				PoolerType: clustermetadatapb.PoolerType_REPLICA,
-				ReplicationStatus: &multipoolermanagerdatapb.StandbyReplicationStatus{
-					LastReceiveLsn:     "0/1234567",
-					WalReceiverStatus:  "streaming",
-					LastMsgReceiveTime: timestamppb.New(now.Add(-5 * time.Second)),
-					PrimaryConnInfo:    &multipoolermanagerdatapb.PrimaryConnInfo{Host: "primary-host", Port: 5432},
-				},
-			},
-		}, nil))
-
-		// New pooler just added by PoolerWatcher: no health data yet
-		// (StreamSnapshotsReceived==0, IsLastCheckValid==false).
-		store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-			MultiPooler: &clustermetadatapb.MultiPooler{
-				Id:       &clustermetadatapb.ID{Component: clustermetadatapb.ID_MULTIPOOLER, Cell: "cell2", Name: "replica2"},
-				ShardKey: &clustermetadatapb.ShardKey{Database: "db1", TableGroup: "tg1", Shard: "shard1"},
-				Type:     clustermetadatapb.PoolerType_REPLICA,
-			},
-			// IsLastCheckValid and StreamSnapshotsReceived are both zero — brand new pooler.
-		}, nil))
-
-		gen := NewAnalysisGenerator(ps, nil)
-		sa, err := gen.GenerateShardAnalysis(&clustermetadatapb.ShardKey{Database: "db1", TableGroup: "tg1", Shard: "shard1"})
-		require.NoError(t, err)
-
-		// replica1 is connected; replica2 has never reported health and must not
-		// count against the connected-replicas check.
-		assert.True(t, replicasStreamingFromLeader(sa),
-			"new pooler with StreamSnapshotsReceived==0 must not count as a disconnected replica")
-	})
-}
-
-func TestPopulatePrimaryInfo_IsInPrimaryStandbyList(t *testing.T) {
-	ps := store.NewTestCache(t)
-
-	primaryID := &clustermetadatapb.ID{
-		Component: clustermetadatapb.ID_MULTIPOOLER,
-		Cell:      "cell1",
-		Name:      "primary-1",
-	}
-
-	replica1ID := &clustermetadatapb.ID{
-		Component: clustermetadatapb.ID_MULTIPOOLER,
-		Cell:      "cell1",
-		Name:      "replica-1",
-	}
-
-	replica2ID := &clustermetadatapb.ID{
-		Component: clustermetadatapb.ID_MULTIPOOLER,
-		Cell:      "cell2",
-		Name:      "replica-2",
-	}
-
-	// Add primary with replica1 in standby list
-	store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-		MultiPooler: &clustermetadatapb.MultiPooler{
-			Id: primaryID,
-			ShardKey: &clustermetadatapb.ShardKey{
-				Database:   "testdb",
-				TableGroup: "testtg",
-				Shard:      "0",
-			},
-			Type: clustermetadatapb.PoolerType_PRIMARY,
-		},
-		IsLastCheckValid: true,
-		LastSeen:         timestamppb.Now(),
-		ConsensusStatus: &clustermetadatapb.ConsensusStatus{
-			Id: primaryID,
-			CurrentPosition: &clustermetadatapb.PoolerPosition{
-				Rule: &clustermetadatapb.ShardRule{
-					RuleNumber:    &clustermetadatapb.RuleNumber{CoordinatorTerm: 1},
-					LeaderId:      primaryID,
-					CohortMembers: []*clustermetadatapb.ID{replica1ID},
-				},
-			},
-		},
-		Status: &multipoolermanagerdatapb.Status{
-			PoolerType:    clustermetadatapb.PoolerType_PRIMARY,
-			PostgresReady: true,
-			PrimaryStatus: &multipoolermanagerdatapb.PrimaryStatus{
-				Lsn:   "0/1234567",
-				Ready: true,
-				SyncReplicationConfig: &multipoolermanagerdatapb.SynchronousReplicationConfiguration{
-					StandbyIds: []*clustermetadatapb.ID{replica1ID},
-				},
-			},
-		},
-	}, nil))
-
-	// Add replica1 (in standby list)
-	store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-		MultiPooler: &clustermetadatapb.MultiPooler{
-			Id: replica1ID,
-			ShardKey: &clustermetadatapb.ShardKey{
-				Database:   "testdb",
-				TableGroup: "testtg",
-				Shard:      "0",
-			},
-			Type: clustermetadatapb.PoolerType_REPLICA,
-		},
-		IsLastCheckValid: true,
-		LastSeen:         timestamppb.Now(),
-		Status: &multipoolermanagerdatapb.Status{
-			PoolerType: clustermetadatapb.PoolerType_REPLICA,
-			ReplicationStatus: &multipoolermanagerdatapb.StandbyReplicationStatus{
-				IsWalReplayPaused: false,
-				Lag:               durationpb.New(100 * time.Millisecond),
-			},
-		},
-	}, nil))
-
-	// Add replica2 (not in standby list)
-	store.SeedCache(t, ps, store.NewPooler(&multiorchdatapb.PoolerHealthState{
-		MultiPooler: &clustermetadatapb.MultiPooler{
-			Id: replica2ID,
-			ShardKey: &clustermetadatapb.ShardKey{
-				Database:   "testdb",
-				TableGroup: "testtg",
-				Shard:      "0",
-			},
-			Type: clustermetadatapb.PoolerType_REPLICA,
-		},
-		IsLastCheckValid: true,
-		LastSeen:         timestamppb.Now(),
-		Status: &multipoolermanagerdatapb.Status{
-			PoolerType: clustermetadatapb.PoolerType_REPLICA,
-			ReplicationStatus: &multipoolermanagerdatapb.StandbyReplicationStatus{
-				IsWalReplayPaused: false,
-				Lag:               durationpb.New(100 * time.Millisecond),
-			},
-		},
-	}, nil))
-
-	generator := NewAnalysisGenerator(ps, nil)
-	shardKey := &clustermetadatapb.ShardKey{Database: "testdb", TableGroup: "testtg", Shard: "0"}
-
-	t.Run("replica in standby list", func(t *testing.T) {
-		sa, err := generator.GenerateShardAnalysis(shardKey)
-		require.NoError(t, err)
-		analysis := findPoolerByName(sa, "replica-1")
-		require.NotNil(t, analysis)
-		assert.True(t, sa.IsInStandbyList(poolerID(analysis)), "replica1 should be in standby list")
-	})
-
-	t.Run("replica not in standby list", func(t *testing.T) {
-		sa, err := generator.GenerateShardAnalysis(shardKey)
-		require.NoError(t, err)
-		analysis := findPoolerByName(sa, "replica-2")
-		require.NotNil(t, analysis)
-		assert.False(t, sa.IsInStandbyList(poolerID(analysis)), "replica2 should not be in standby list")
-	})
 }
 
 // TestPopulatePrimaryInfo_PicksHighestPrimaryTerm verifies that when two primaries transiently
